+++ ui-src/js/system-controller.js
@@ -2,7 +2,7 @@ angular
     .module('nzbhydraApp')
     .controller('SystemController', SystemController);
 
-function SystemController($scope, $state, growl, RestartService, NzbHydraControlService) {
+function SystemController($scope, $state, $http, growl, RestartService, NzbHydraControlService) {
 
 
     $scope.shutdown = function () {
@@ -56,7 +56,20 @@ function SystemController($scope, $state, growl, RestartService, NzbHydraControl
 
     $scope.goToState = function (index) {
         $state.go($scope.tabs[index].state);
-    }
     
+    };
+
+    $scope.downloadDebuggingInfos = function() {
+        $http({method: 'GET', url: '/internalapi/getdebugginginfos', responseType: 'arraybuffer'}).success(function (data, status, headers, config) {
+            var a = document.createElement('a');
+            console.log(data);
+            var blob = new Blob([data], {'type': "application/octet-stream"});
+            a.href = URL.createObjectURL(blob);
+            a.download = "filename.zip";
+            a.click();
+        }).error(function (data, status, headers, config) {
+            // handle error
+        });
+    }
     
 }