+++ changelog.md
@@ -4,6 +4,8 @@
 ### 0.2.106
 Fixed: Forbidden and ignored words not displayed in config.
 
+Fixed: Download of bugreport would not work with auth type "Form".
+
 ### 0.2.105
 Fixed: Set appropriate categories when API searches are done ("Movies" for movie search, etc.)